requirements.txt
langchain
langgraph
langchain_community
langchain_groq
python-dotenv
fastapi
uvicorn
requests 